Crate nanomsg_sys

Reexports

pub use libc::*;

pub use posix_consts::*;

Modules

posix_consts

Structs

nn_pollfd

Constants

AF_SP

AF_SP_RAW

EFSM

ETERM

NN_BUS

NN_DOMAIN

NN_DONTWAIT

NN_INPROC

NN_IPC

NN_IPV4ONLY

NN_LINGER

NN_MSG

NN_PAIR

NN_POLLIN

NN_POLLOUT

NN_POLL_IN_AND_OUT

NN_PROTOCOL

NN_PROTO_BUS

NN_PROTO_PAIR

NN_PROTO_PIPELINE

NN_PROTO_PUBSUB

NN_PROTO_REQREP

NN_PROTO_SURVEY

NN_PUB

NN_PULL

NN_PUSH

NN_RCVBUF

NN_RCVFD

NN_RCVPRIO

NN_RCVTIMEO

NN_RECONNECT_IVL

NN_RECONNECT_IVL_MAX

NN_REP

NN_REQ

NN_REQ_RESEND_IVL

NN_RESPONDENT

NN_SNDBUF

NN_SNDFD

NN_SNDPRIO

NN_SNDTIMEO

NN_SOCKADDR_MAX

NN_SOCKET_NAME

NN_SOL_SOCKET

NN_SUB

NN_SUB_SUBSCRIBE

NN_SUB_UNSUBSCRIBE

NN_SURVEYOR

NN_SURVEYOR_DEADLINE

NN_TCP

NN_TCP_NODELAY

Functions

nn_allocmsg
http://nanomsg.org/v0.4/nn_allocmsg.3.html

nn_bind
"Adds a local endpoint to the socket s. The endpoint can be then used by other applications to connect to. The addr argument consists of two parts as follows: transport://address. The transport specifies the underlying transport protocol to use. The meaning of the address part is specific to the underlying transport protocol."

nn_close
"Closes the socket s. Any buffered inbound messages that were not yet received by the application will be discarded. The library will try to deliver any outstanding outbound messages for the time specified by NN_LINGER socket option. The call will block in the meantime."

nn_connect
"Adds a remote endpoint to the socket s. The library would then try to connect to the specified remote endpoint. The addr argument consists of two parts as follows: transport://address. The transport specifies the underlying transport protocol to use. The meaning of the address part is specific to the underlying transport protocol."

nn_device
http://nanomsg.org/v0.4/nn_device.3.html

nn_errno
http://nanomsg.org/v0.4/nn_errno.3.html

nn_freemsg
http://nanomsg.org/v0.4/nn_freemsg.3.html

nn_getsockopt
"Retrieves the value for the option. The level argument specifies the protocol level at which the option resides. For generic socket-level options use NN_SOL_SOCKET level. For socket-type-specific options use socket type for level argument (e.g. NN_SUB). For transport-specific options use ID of the transport as the level argument (e.g. NN_TCP).

nn_poll
http://nanomsg.org/v0.4/nn_poll.3.html

nn_reallocmsg
http://nanomsg.org/v0.4/nn_reallocmsg.3.html

nn_recv
"Receive a message from the socket s and store it in the buffer referenced by the buf argument. Any bytes exceeding the length specified by the len argument will be truncated.

nn_recvmsg
http://nanomsg.org/v0.4/nn_recvmsg.3.html

nn_send
"The function will send a message containing the data from buffer pointed to by buf parameter to the socket s. The message will be len bytes long. Alternatively, to send a buffer allocated by nn_allocmsg(3) function set the buf parameter to point to the pointer to the buffer and len parameter to NN_MSG constant. In this case a successful call to nn_send will deallocate the buffer. Trying to deallocate it afterwards will result in undefined behaviour.

nn_sendmsg
http://nanomsg.org/v0.4/nn_sendmsg.3.html

nn_setsockopt
"Sets the value of the option. The level argument specifies the protocol level at which the option resides. For generic socket-level options use NN_SOL_SOCKET level. For socket-type-specific options use socket type for level argument (e.g. NN_SUB). For transport-specific options use ID of the transport as the level argument (e.g. NN_TCP).

nn_shutdown
"Removes an endpoint from socket s. how parameter specifies the ID of the endpoint to remove as returned by prior call to nn_bind(3) or nn_connect(3). nn_shutdown() call will return immediately, however, the library will try to deliver any outstanding outbound messages to the endpoint for the time specified by NN_LINGER socket option."

nn_socket
"Creates an SP socket with specified domain and protocol. Returns a file descriptor for the newly created socket."

nn_strerror
http://nanomsg.org/v0.4/nn_strerror.3.html

nn_symbol

nn_term
http://nanomsg.org/v0.4/nn_term.3.html